Implement new admin get self endpoint, move auth store to the portal, and implement working password changing and name changing
