+++ packages/builder/src/pages/builder/portal/apps/index.svelte
@@ -17,8 +17,7 @@
   import api, { del } from "builderStore/api"
   import analytics from "analytics"
   import { onMount } from "svelte"
-  import { apps } from "stores/portal"
-  import { auth } from "stores/backend"
+  import { apps, auth } from "stores/portal"
   import download from "downloadjs"
   import { goto } from "@roxi/routify"
   import ConfirmDialog from "components/common/ConfirmDialog.svelte"
