+++ packages/builder/src/stores/backend/index.js
@@ -7,4 +7,3 @@ export { roles } from "./roles"
 export { datasources } from "./datasources"
 export { integrations } from "./integrations"
 export { queries } from "./queries"
-export { auth } from "./auth"
