+++ packages/builder/src/stores/portal/auth.js
@@ -7,7 +7,7 @@ export function createAuthStore() {
   return {
     subscribe: store.subscribe,
     checkAuth: async () => {
-      const response = await api.get("/api/self")
+      const response = await api.get("/api/admin/users/self")
       const user = await response.json()
       if (response.status === 200) {
         store.update(state => ({ ...state, user }))
@@ -32,6 +32,14 @@ export function createAuthStore() {
       }
       await response.json()
       store.update(state => ({ ...state, user: null }))
+    },
+    updateSelf: async user => {
+      const response = await api.post("/api/admin/users/self", user)
+      if (response.status === 200) {
+        store.update(state => ({ ...state, user: { ...state.user, ...user } }))
+      } else {
+        throw "Unable to update user details"
+      }
     },
     forgotPassword: async email => {
       const response = await api.post(`/api/admin/auth/reset`, {
