+++ packages/builder/src/stores/portal/index.js
@@ -3,3 +3,4 @@ export { users } from "./users"
 export { admin } from "./admin"
 export { apps } from "./apps"
 export { email } from "./email"
+export { auth } from "./auth"
